+++ CHANGELOG.md
@@ -7,7 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
 - hnefatafl-client: Hide the password.
+- hnefatafl-client: do logging.
 
 ## [0.4.1] - 2025-02-04